remove MainApp.tsx
